Remove gadget name link target to metawiki from table output
Linking each gadget to its metawiki page (existing or not) could make
users think that metawiki hosts the canonical source of that gadget's
code or wonder why in some cases the link is red when that gadget is
not available on metawiki.
There is no good way to identify the canonical source location of a
gadget, hence remove the metawiki link target from the first column.
Users can still investigate by using the links in the second column.

See https://phabricator.wikimedia.org/T251362 for discussion.